requirements.txt
db-dtypes==1.3.0
google-auth==2.35.0
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.27.0
google-cloud-storage==2.18.2
pandas==2.2.3
pyarrow==18.0.0
